feat: finish button and spinner
