+++ src/components/molecules/Clickable/Clickable.tsx
@@ -54,7 +54,7 @@ const colors = (type: Type): string => {
 
   const options: Record<Type, string> = {
     primary:
-      "bg-gradient-to-tr from-black to-gray-800 text-gray-50 hover:from-gray-100 hover:to-white hover:text-black",
+      "bg-gradient-to-tr from-black to-gray-800 text-gray-50 hover:from-gray-100 hover:to-white hover:text-black border border-0 border-gray-200 hover:border-1",
     secondary: "bg-transparent border border-gray-100 text-gray-900 hover:border-gray-400 hover:text-black",
     alert: "bg-gradient-to-tr from-red-600 to-red-500 text-white hover:to-red-400",
   }
