+++ src/components/molecules/Clickable/Link.stories.mdx
@@ -1,0 +1,17 @@
+import { Meta, Story, Canvas, Preview, ArgsTable } from "@storybook/addon-docs/blocks"
+
+import { Link } from "./Link"
+
+<Meta title="molecules/Link" component={Link} />
+
+# Link
+
+<Preview>
+  <Story name="standard">
+    <Link size="medium" label="Link" href="#" />
+  </Story>
+</Preview>
+
+## Arguments
+
+<ArgsTable of={Link} />
