+++ tsconfig.json
@@ -15,5 +15,5 @@
     "jsx": "preserve"
   },
   "exclude": ["node_modules", "dist"],
-  "include": ["**/*.ts", "**/*.tsx", "**/*.js", ".storybook/storybook.test.ts"]
+  "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.mdx", ".storybook/storybook.test.ts"]
 }
